ext/requirements.txt
absl-py==0.1.13
astor==0.6.2
backcall==0.1.0
bleach==1.5.0
certifi==2018.1.18
chardet==3.0.4
cycler==0.10.0
decorator==4.2.1
easydict==1.7
entrypoints==0.2.3
gast==0.2.0
graphviz==0.8.2
grpcio==1.10.1
html5lib==0.9999999
idna==2.6
ipykernel==4.8.2
ipython==6.3.1
ipython-genutils==0.2.0
ipywidgets==7.2.1
jedi==0.11.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
kiwisolver==1.0.1
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.2.2
mistune==0.8.3
mxnet-cu90==1.1.0.post0
mxnet-cu91==1.1.0.post0
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.1
notebook==5.4.1
numpy==1.14.2
opencv-python==3.4.0.12
pandocfilters==1.4.2
parso==0.1.1
pexpect==4.4.0
pickleshare==0.7.4
Pillow==5.1.0
prompt-toolkit==1.0.15
protobuf==3.5.2.post1
ptyprocess==0.5.2
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.7.2
pytz==2018.4
PyWavelets==0.5.2
pyzmq==17.0.0
qtconsole==4.3.1
requests==2.18.4
scikit-image==0.13.1
scikit-learn==0.19.1
scipy==1.0.1
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
tensorboard==1.6.0
tensorflow-gpu==1.5.0
tensorflow-tensorboard==1.5.1
tensorlayer==1.8.4
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
toposort==1.5
tornado==5.0.2
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
Werkzeug==0.14.1
widgetsnbextension==3.2.1
